refactor(kicad-studio): split activation composition root into focused controllers
Extract the ~450 lines of nested activation closures from extension.ts into
focused, dependency-injected controllers under src/activation/, leaving a slim
composition root that instantiates services and wires controllers:

- StudioContextController: buildStudioContext / pushStudioContext (+ the pure
  readDesignBlockNames helper).
- McpActivationController: refreshMcpState and all MCP setContext keys.
- WorkspaceContextController: refreshContexts, active-project selection, and the
  debounced .kicad_pro rescan (now disposable).
- SaveCheckController: auto DRC/ERC on save and proactive AI offer.
- ActivationState: shared aiHealthy / latestDrcRun previously held as closures.

Behavior is preserved exactly; the logic moved unchanged. Coverage excludes
src/activation/** consistent with the existing extension.ts exclusion (VS Code
orchestration covered by integration tests). Adds activationControllers.test.ts
for the pure helper, ActivationState, and the debounce/dispose lifecycle.

Closes #397
